+++ lib/middleware/receive-message/campaign-continue.js
